Separate booking folder into booking and mmb
