+++ apps/graphql/src/Schema.js
@@ -5,14 +5,14 @@ import { GraphQLSchema, GraphQLObjectType } from 'graphql';
 import SearchOneWayItineraries from './apps/itinerary/queries/SearchOneWayItineraries';
 import SearchReturnItineraries from './apps/itinerary/queries/SearchReturnItineraries';
 import LocationsByTerm from './apps/location/queries/LocationsByTerm';
-import CustomerBookings from './apps/booking/queries/CustomerBookings';
-import BookingDetail from './apps/booking/queries/BookingDetail';
+import CustomerBookings from './apps/manageMyBooking/queries/CustomerBookings';
+import BookingDetail from './apps/manageMyBooking/queries/BookingDetail';
 import SaveBooking from './apps/booking/mutations/SaveBooking';
 import ConfirmPayment from './apps/booking/mutations/ConfirmPayment';
 import CheckItinerary from './apps/itinerary/queries/CheckItinerary';
-import BookingOneWay from './apps/booking/types/outputs/BookingOneWay';
-import BookingMulticity from './apps/booking/types/outputs/BookingMulticity';
-import BookingReturn from './apps/booking/types/outputs/BookingReturn';
+import BookingOneWay from './apps/manageMyBooking/types/outputs/BookingOneWay';
+import BookingMulticity from './apps/manageMyBooking/types/outputs/BookingMulticity';
+import BookingReturn from './apps/manageMyBooking/types/outputs/BookingReturn';
 import GeoIP from './apps/geoIp/queries/GeoIP';
 import ItineraryReturn from './apps/itinerary/types/outputs/ItineraryReturn';
 import ItineraryOneWay from './apps/itinerary/types/outputs/ItineraryOneWay';
