+++ apps/graphql/src/apps/manageMyBooking/queries/__tests__/__fixtures__/BookingDetail.graphql
