+++ apps/graphql/src/apps/manageMyBooking/queries/__tests__/__fixtures__/CustomerBookings.graphql
