+++ apps/graphql/src/apps/manageMyBooking/types/outputs/BagType.js
