+++ apps/graphql/src/apps/manageMyBooking/types/outputs/BookingOneWay.js
@@ -3,29 +3,24 @@
 import { GraphQLObjectType } from 'graphql';
 
 import BookingInterface, { commonFields } from './BookingInterface';
-import type { Booking } from '../../Booking';
+import type { Booking } from '../../manageMyBooking';
 import GraphQLSector from '../../../common/types/outputs/Sector';
 import FromToInterface from '../../../common/types/outputs/FromToInterface';
 
-const NAME = 'BookingReturn';
+const NAME = 'BookingOneWay';
 
-const BookingReturn = new GraphQLObjectType({
+const BookingOneWay = new GraphQLObjectType({
   name: NAME,
-  description: 'Booking from A to B with return, and possible stopovers',
+  description: 'Booking from A to B with no return, but possible stopovers',
   interfaces: [BookingInterface, FromToInterface],
   isTypeOf: (value: Booking) => value.type === NAME,
   fields: {
     ...commonFields,
-
-    inbound: {
+    sector: {
       type: GraphQLSector,
-      resolve: ({ inbound }: Booking) => inbound,
-    },
-    outbound: {
-      type: GraphQLSector,
-      resolve: ({ outbound }: Booking) => outbound,
+      resolve: (booking: Booking) => booking,
     },
   },
 });
 
-export default BookingReturn;
+export default BookingOneWay;
