+++ apps/graphql/src/apps/manageMyBooking/types/outputs/BookingReturn.js
@@ -3,24 +3,29 @@
 import { GraphQLObjectType } from 'graphql';
 
 import BookingInterface, { commonFields } from './BookingInterface';
-import type { Booking } from '../../Booking';
+import type { Booking } from '../../manageMyBooking';
 import GraphQLSector from '../../../common/types/outputs/Sector';
 import FromToInterface from '../../../common/types/outputs/FromToInterface';
 
-const NAME = 'BookingOneWay';
+const NAME = 'BookingReturn';
 
-const BookingOneWay = new GraphQLObjectType({
+const BookingReturn = new GraphQLObjectType({
   name: NAME,
-  description: 'Booking from A to B with no return, but possible stopovers',
+  description: 'Booking from A to B with return, and possible stopovers',
   interfaces: [BookingInterface, FromToInterface],
   isTypeOf: (value: Booking) => value.type === NAME,
   fields: {
     ...commonFields,
-    sector: {
+
+    inbound: {
       type: GraphQLSector,
-      resolve: (booking: Booking) => booking,
+      resolve: ({ inbound }: Booking) => inbound,
+    },
+    outbound: {
+      type: GraphQLSector,
+      resolve: ({ outbound }: Booking) => outbound,
     },
   },
 });
 
-export default BookingOneWay;
+export default BookingReturn;
